Add "TaskBook"
1. User can open "TaskBook" and edit by double-clicking the name;
2. In "TaskBook", user can add anything related to the task. The related items will be shown in the "All task list";
3. User can add "specific task" from "all task list"; ("spcific task" are specific procedure to finish the task)
4. User can add "subtask" from "all task list"; ("subtask" are more specific procedure to finish the "specific task").

+++ TnP/TnP.Model/Elements/EPlan.cs
@@ -26,10 +26,10 @@ namespace TnP.Model.Elements
         public bool endState;//finished or gave up
         public DateTime endT;//record the time a plan finally over
 
-        public int lplanID;//belong to which long term plan
-        public int taskID;//belong to which task
-        public int goalID;//belong to which goal
-        public int thoughtID;//belong to which thought
+        public long lplanID;//belong to which long term plan
+        public long taskID;//belong to which task
+        public long goalID;//belong to which goal
+        public long thoughtID;//belong to which thought
 
         public Feature feature;//record the characteristic of a plan
         public List<Step> steps = new List<Step>();//record the steps of a plan
